requirements.txt
torch
torchvision
botorch


bayesian-optimization==1.2.0
matplotlib==3.3.4
numpy
pandas
pyaml==20.4.0
pyyaml==5.4.1
scikit-learn==0.23.2
scikit-optimize==0.8.1
scipy==1.4.1
seaborn==0.11.1
stringcase==1.2.0
tqdm==4.56.0

